Switch Node.js shell's stdin to raw mode
In the normal "cooked" mode, Node only produces data when <return> is
pressed, giving us whatever had been typed up to that point. This means
that the arrow keys would print their control sequences as part of the
input string, instead of readline receiving them individually.

By enabling raw mode, we receive each key as it happens and can respond
to it, meaning that the arrow keys work as expected. We have to change
how we detect SIGINT and SIGQUIT though, since Node will no longer
produce the event we previously were listening to.

Fixes #20.