Add a basic test harness for builtin coreutil functions
Pass MakeTestContext the command, and the arguments/options you want.
You can then run the command using that context, and examine the return
value, any thrown error, and the contents of stdout and stderr. This
does not yet let us test builtins that access the filesystem.

Tests for `true` and `false` are included as very basic examples.