Print command help with `help COMMAND`
`help COMMAND` now looks to see if `COMMAND` is a builtin, and if so,
prints some documentation for it: Usage, a description, and a list of
possible options.

`command.usage` can be a string or an array of strings, if there are
multiple ways a command can be used. If it's not specified, we generate
one with the form `foo [OPTIONS] INPUTS...`, depending on whether any
options or positional arguments are allowed.

`command.description` is a string. If it's not specified, we print
nothing.

For each option, we print the short name, long name, value name, and
description, depending on which are present. The valueName is used when
the option is not a boolean, defaults to `VALUE`, and is printed as
`--option=VALUE`.

This is intentionally put in its own file, so we can eventually call
`printUsage()` when running `foo --help` as well as `help foo`.