Use lstat instead of stat in Node filesystem layer
lstat is identical to stat, except that if the target is a symbolic
link, it returns the information about the link, not the target file.
This is what we actually want.

This fixes an error where `readdir()` of a directory that contains a
link to a non-existing file would throw ENOENT. We don't care if that
file exists (though it would be nice to indicate that through colour).