Resolve ~ in paths as the home directory
We already had code for this in `ls`, but it was buggy: removing the
first character of `~/foo` produces `/foo`, which then always resolves
to `/foo` regardless of what parent path you give it.

This fixes that by prepending a `.`, makes this a function in a new
`utils/path.js` file, and then uses that in every location I could find
that tries to resolve relative paths.