Make sleep builtin test more tolerant of minor variations
This was occasionally failing for me, and on CI. Instead of requiring
that it's >= the requested duration, ensure it's within a few
milliseconds. This allows for pauses that are very slightly less than
requested, and also prevents them being massively too long. Report the
expected and actual durations if it fails, for good measure.