On Node, send console.foo() logging to a file, or discard it
On Node, having this console output appear inline makes it difficult to
check that normal output is correct, and also looks confusing for
non-developers. Instead, intercept calls to `console.foo()`, and either
write the logs to a file as indicated by `--log=FILE`, or discard them.
Logging can still be followed easily by running `tail -f FILE` in a
separate terminal.

For now, all kinds of logs are sent to the same file. If necessary in
future, we could direct different types to different files.