Implement a 'which' builtin
CommandProviders can expose this by setting the 'path' field of the
object they return from execute(). If 'path' is not set, we assume that
the command is a shell built-in.

Scripts set their 'path' to be the same as the path used to access them,
which matches the behaviour I'm seeing on Linux.